+++ app/models/backup_library.rb
@@ -1,0 +1,2 @@
+class BackupLibrary < ActiveRecord::Base
+end