+++ app/models/backup_location.rb
@@ -1,0 +1,3 @@
+class BackupLocation < ActiveRecord::Base
+  belongs_to :backup_library
+end