+++ CHANGELOG.md
@@ -13,7 +13,11 @@ and releases in Discovery project adheres to [Semantic Versioning](http://semver
 - data sanity regression test values for databases [PR#1915](https://github.com/ualbertalib/discovery/pull/1915)
 - hours and wording of question for BPSC request form [#1889](https://github.com/ualbertalib/discovery/issues/1889)
 
+### Added
+- create backup tables for library and location [#1843](https://github.com/ualbertalib/discovery/issues/1843)
+
 ## [3.0.116] - 2020-02-02
+
 ### Added
 - Added "See License Terms" link to SFX e-journal record holdings. [#1893]